refactor: split ui managers into modules
