+++ src/progressView/index.html
@@ -31,7 +31,11 @@
           "./modules/formatters.js": "${formattersUri}",
           "./modules/taskManagers.js": "${taskManagersUri}",
           "./modules/usageManagers.js": "${usageManagersUri}",
-          "./modules/uiManagers.js": "${uiManagersUri}",
+          "./modules/uiManagers/StreamTabs.js": "${streamTabsUri}",
+          "./modules/uiManagers/Toolbar.js": "${toolbarUri}",
+          "./modules/uiManagers/Status.js": "${statusUri}",
+          "./modules/uiManagers/FileList.js": "${fileListUri}",
+          "./modules/uiManagers/Events.js": "${eventsUri}",
           "./modules/katexMacros.js": "${katexMacrosUri}",
           "@common/webviewContext.js": "${webviewContextUri}",
           "@common/webviewState.js": "${webviewStateUri}",
